editoast: add CanRead check for level_crossing_occupancy endpoint
